move common charting funcs to plotting and add query for adminStats
